requirements.txt
absl-py==1.4.0
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==4.1.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.34.1
h5py==3.1.0
idna==3.7
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.12.2
ipywidgets==8.1.2
jedi==0.18.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.3.4
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.82
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==1.2.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydot==2.0.0
Pygments==2.15.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
requests==2.32.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.0
scipy==1.6.0
seaborn==0.11.0
Send2Trash==1.8.2
setuptools==69.5.1
sip==6.7.12
six==1.15.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-model-optimization==0.8.0
termcolor==1.1.0
terminado==0.17.1
tf-keras==2.15.0
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.3
typing-extensions==3.7.4.3
urllib3==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
wrapt==1.12.1
zipp==3.17.0
